% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeinfo.r
\name{codeinfo}
\alias{codeinfo}
\title{Information on the system, R-version, packages, files and R-objects, for traceability}
\usage{
codeinfo(r = TRUE, packages = character(0), files = character(0),
  files_path = character(0), objects = list())
}
\arguments{
\item{r}{Single logical value. If \code{TRUE} (the default), 
the function returns information on the system and R.}

\item{packages}{Vector of character strings, naming installed packages 
for which information should be returned. If 
\code{length(packages)} is \code{0} (the default), no 
information is returned on packages.}

\item{files}{Vector of character strings. Name, optionally including 
the full or relative path (see \code{files_path} below), 
for which \code{\link[tools:md5sum]{MD5 Checksums}} should 
be returned. If \code{length(files)} is \code{0} 
(the default), no information is returned on files.}

\item{files_path}{Single character string. General path to a single folder 
where all the files in \code{files} are located. If the 
path differ between files, then it should be passed via 
\code{files} instead. Default value for \code{files} is 
\code{character(0)}, meaning that no general path is 
provided, and that the files are either in the working 
directory or included in \code{files} instead.}

\item{objects}{A labelled list of arbitrary R-objects, whose 
\code{\link[tools:md5sum]{MD5 Checksums}} should be 
returned. For example \code{objects =}\code{list( a = object1, } 
\code{b = object2}, where \code{object1} and \code{object2} 
are R-objects and \code{a} and \code{b} are labels.}
}
\description{
Information on the system, R-version, packages, files and 
 R-objects, for traceability. Includes checksums of 
 files, packages and R-objects.
}
\examples{

library( "codeinfo" )

#   Create an arbitrary R object to be "hashed"
m <- matrix( 1:4, 2, 2 )

#   Create an arbitrary file for which the MD5 checksum will
#   be calculated
tf1 <- tempfile() 
tf2 <- tempfile() 
writeLines( text = c( letters, LETTERS, 0:9 ), con = tf1 )
writeLines( text = c( letters, LETTERS, 0:8 ), con = tf2 )



#   use codeinfo() to obtain information
codeinfo(
    packages = c( "utils", "tools" ), 
    files      = c( tf1, tf2 ), 
    objects = list( "m" = m, "quantile" = quantile ) )



#   Clean-up
file.remove( tf1, tf2 )
rm( m, tf1, tf2 )
}
\seealso{
\code{\link{Sys.info}}, 
 \code{\link[utils]{sessionInfo}}, 
 \code{\link[utils]{installed.packages}} and 
 \code{\link[tools]{md5sum}}.
}
