% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeinfo.r
\name{package_md5sum}
\alias{package_md5sum}
\title{Compute the MD5 checksums of each files in an installed package and the checksum of the checksums}
\usage{
package_md5sum(package, exclude_MD5 = TRUE)
}
\arguments{
\item{package}{Single character string. Name of an installed package.}

\item{exclude_MD5}{Single logical value. If \code{TRUE} (the default), the 
file MD5 is excluded from the output.}
}
\value{
Returns a \code{\link[base]{list}} with two named items: 
 \code{checksums} and \code{checksums_checksum}. 
 \code{checksums} is a \code{\link[base]{data.frame}} with 
 the \code{\link[tools:md5sum]{MD5 checksums}} of each files 
 in the package (the first column contains the checksums 
 and the second column the name and relative path to the 
 file). It is formatted to look alike the \code{MD5} file 
 of installed packages (when present). 
 \code{checksums_checksum} is a single character string, 
 the MD5 checksum calculated on a file containing 
 the table \code{checksums}, without row and column 
 names, with a space as column separator and without quotes.
}
\description{
Compute the \code{\link[tools:md5sum]{MD5 checksums}} of 
 each files in an installed package and the checksum of 
 the checksums (i.e. the MD5 checksum of a file containing 
 the checksums of each files in the package)
}
\seealso{
\code{\link[base]{find.package}} and 
 \code{\link[tools]{md5sum}}.
}
